fix: 권한이 없는 사용자 접근 예외처리 (#86)
