+++ src/main/java/com/api/TaveShot/global/config/SecurityConfig.java
@@ -2,6 +2,8 @@ package com.api.TaveShot.global.config;
 
 import com.api.TaveShot.domain.Member.domain.Role;
 import com.api.TaveShot.global.security.jwt.JwtAuthenticationFilter;
+import com.api.TaveShot.global.security.oauth2.CustomAccessDeniedHandler;
+import com.api.TaveShot.global.security.oauth2.CustomAuthenticationEntryPoint;
 import com.api.TaveShot.global.security.oauth2.CustomOAuth2UserService;
 import com.api.TaveShot.global.security.oauth2.CustomOAuthSuccessHandler;
 import lombok.RequiredArgsConstructor;
@@ -26,6 +28,8 @@ public class SecurityConfig {
     private final CustomOAuth2UserService customOAuth2UserService;
     private final JwtAuthenticationFilter jwtAuthenticationFilter;
     private final CustomOAuthSuccessHandler customOAuthSuccessHandler;
+    private final CustomAuthenticationEntryPoint customAuthEntryPoint;
+    private final CustomAccessDeniedHandler customAccessDeniedHandler;
 
     @Bean
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
@@ -63,7 +67,11 @@ public class SecurityConfig {
                 .userInfoEndpoint()
                 .userService(customOAuth2UserService)
                     .and()
-                .successHandler(customOAuthSuccessHandler);
+                .successHandler(customOAuthSuccessHandler)
+                    .and()
+                .exceptionHandling()
+                        .authenticationEntryPoint(customAuthEntryPoint)
+                        .accessDeniedHandler(customAccessDeniedHandler);
         http.addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class);
 
         return http.build();
