+++ src/main/java/com/api/TaveShot/global/exception/ErrorType.java
@@ -87,8 +87,9 @@ public enum ErrorType {
     // ------------------------------------------ Compile ------------------------------------------
     _COMPILE_PROBLEM_NOT_FOUND(NOT_FOUND, "PROBLEM_4040", "요청한 문제 번호를 찾을 수 없습니다."),
     _SUBMIT_PAGE_NOT_FOUND(NOT_FOUND, "COMPILE_4040", "아직 정답이 준비 되지 않아 코드를 제출할 수 없습니다."),
-    _PROBLEM_CONVERSION_ERROR(INTERNAL_SERVER_ERROR, "PROBLEM_5002", "문제 정보 변환 중 오류가 발생했습니다.")
+    _PROBLEM_CONVERSION_ERROR(INTERNAL_SERVER_ERROR, "PROBLEM_5002", "문제 정보 변환 중 오류가 발생했습니다."),
 
+    _ACCESS_DENIED(FORBIDDEN, "ACCESS_4031", "권한이 없습니다."),
     ;
 
     private final HttpStatus status;
